chore: extract BlendShapePreviewContext from RemoveMeshWithBoxPreviewContext
